feat: edit profile page (#58)
* refactor: separate register to a component

* feat: fill in form data

* feat: edit profile
