+++ src/components/Profile/Register/index.ts
@@ -1,0 +1,3 @@
+export * from './constants';
+export { getFormData } from './getFormData';
+export { RegisterForm } from './RegisterForm';
