+++ src/dto/userDTO.ts
@@ -9,6 +9,8 @@ export interface UserDTO {
     nickname?: string;
     email?: string;
     phone?: string;
+    emer_phone?: string;
+    emer_relation?: string;
     line_id?: string;
     disease?: string;
     allergy_food?: string;
@@ -20,4 +22,5 @@ export interface UserDTO {
     group_id?: string;
     baan_id?: string;
     is_got_ticket?: boolean;
+    want_bottle?: boolean;
 }
